requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-log==0.4.0
colorama==0.4.6
Crawl4AI==0.7.4
cryptography==46.0.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.119.0
fastuuid==0.13.5
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanize==4.13.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.0
litellm==1.78.0
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.4
llama-index-cli==0.5.3
llama-index-core==0.14.4
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.4
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-postgres==0.7.0
llama-index-vector-stores-qdrant==0.8.6
llama-index-workflows==2.8.0
llama-parse==0.6.54
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
openai==1.109.1
packaging==25.0
pandas==2.2.3
patchright==1.55.2
pgvector==0.4.1
pillow==11.3.0
platformdirs==4.5.0
playwright==1.55.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rtree==1.4.1
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
striprtf==0.0.26
tenacity==9.1.2
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
trimesh==4.8.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
pymongo==4.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
